Fix stats and dates
