+++ api/api/controllers/cluster_controller.py
@@ -15,7 +15,8 @@ import wazuh.stats as stats
 from api.authentication import get_permissions
 from api.models.base_model_ import Data
 from api.util import remove_nones_to_dict, exception_handler, parse_api_param, raise_if_exc
-from wazuh.core.cluster.dapi.dapi import DistributedAPI, get_system_nodes
+from wazuh.core.cluster.dapi.dapi import DistributedAPI
+from wazuh.core.cluster.control import get_system_nodes
 from wazuh.exception import WazuhError
 
 loop = asyncio.get_event_loop()
@@ -268,18 +269,20 @@ def get_stats_node(node_id, pretty=False, wait_for_complete=False, date=None):
     :param date: Selects the date for getting the statistical information. Format YYYY-MM-DD.
     """
     if date:
-        year, month, day = date.split('-')
+        try:
+            date = datetime.datetime.strptime(date, '%Y-%m-%d')
+            today = False
+        except ValueError:
+            raise WazuhError(1301)
     else:
-        today = datetime.datetime.now()
-        year = str(today.year)
-        month = str(today.month)
-        day = str(today.day)
+        date = datetime.datetime.today()
+        today = True
 
     f_kwargs = {'node_id': node_id,
-                'year': year,
-                'month': month,
-                'day': day,
-                'date': True if date else False}
+                'year': date.year,
+                'month': date.month,
+                'day': date.day,
+                'today': today}
 
     nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=stats.totals,
