+++ framework/wazuh/core/cluster/control.py
@@ -3,7 +3,7 @@
 # This program is free software; you can redistribute it and/or modify it under the terms of GPLv2
 import json
 
-from wazuh import common
+from wazuh import common, WazuhInternalError
 from wazuh.core.cluster import local_client
 from wazuh.core.cluster.common import as_wazuh_object, WazuhJSONEncoder
 from wazuh.core.core_agent import Agent
@@ -89,3 +89,16 @@ async def get_agents(lc: local_client.LocalClient, filter_node=None, filter_stat
     filled_result = [{**r, **{key: 'unknown' for key in select_fields - r.keys()}} for r in result['items']]
     result['items'] = filled_result
     return result
+
+
+async def get_system_nodes():
+    try:
+        lc = local_client.LocalClient()
+        try:
+            result = await get_nodes(lc)
+        finally:
+            lc.transport.close()
+
+        return [node['name'] for node in result['items']]
+    except Exception:
+        raise WazuhInternalError(3012)
