+++ framework/wazuh/exception.py
@@ -117,6 +117,9 @@ class WazuhException(Exception):
                },
 
         # Stats: 1300 - 1399
+        1301: {'message': 'Invalid date',
+               'remediation': 'Please, make sure you use a valid date value)'
+               },
         1307: {'message': 'Invalid parameters',
                'remediation': 'Please, check that the update is correct, there is a problem while reading the results, contact us at [official repository](https://github.com/wazuh/wazuh/issues)'
                },
